Migrate to boost process v2 (#338)
Migrate to boost process v2, fixes compile problems on newer versions of boost.

In addition:
- fix autowrap API breakage
- fix doc build
- fix concurrency issue in keyvi index which appeared after boost process update on Mac

+++ python/src/addons/JsonVector.pyx
@@ -1,7 +1,7 @@
 
 
     def __getitem__(self,  index ):
-        assert isinstance(index, (int, long)), 'arg index wrong type'
+        assert isinstance(index, (int, int)), 'arg index wrong type'
 
         cdef libcpp_utf8_string _r = self.inst.get().Get((<size_t>index))
         py_result = json.loads(_r.decode('utf-8'))
